domjudje files

+++ Skeletons/industrial revolution (transport)/transport2.py
@@ -6,7 +6,7 @@ You can assume that if a shipment arrives, it arrives with either all of its car
 """
 
 
-def get_shipment_loss(file_incoming: str, file_outgoing: str) -> int:
+def get_shipment_loss(file_incoming: str, file_outgoing: str) -> float:
     return 0
